# server/openapi_server/models/data_set.py
# coding: utf-8

from __future__ import absolute_import

from typing import List  # noqa: F401

from openapi_server import util
from openapi_server.models.variable_presentation import VariablePresentation
from openapi_server.models.base_model_ import Model


class DataSet(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, label=None, has_format=None, description=None, has_dimensionality=None, has_presentation=None, type=None, id=None):  # noqa: E501
        """DataSet - a model defined in OpenAPI

        :param label: The label of this DataSet.  # noqa: E501
        :type label: str
        :param has_format: The has_format of this DataSet.  # noqa: E501
        :type has_format: str
        :param description: The description of this DataSet.  # noqa: E501
        :type description: str
        :param has_dimensionality: The has_dimensionality of this DataSet.  # noqa: E501
        :type has_dimensionality: int
        :param has_presentation: The has_presentation of this DataSet.  # noqa: E501
        :type has_presentation: List[VariablePresentation]
        :param type: The type of this DataSet.  # noqa: E501
        :type type: List[str]
        :param id: The id of this DataSet.  # noqa: E501
        :type id: str
        """
        self.openapi_types = {
            'label': str,
            'has_format': str,
            'description': str,
            'has_dimensionality': int,
            'has_presentation': List[VariablePresentation],
            'type': List[str],
            'id': str
        }

        self.attribute_map = {
            'label': 'label',
            'has_format': 'has_format',
            'description': 'description',
            'has_dimensionality': 'has_dimensionality',
            'has_presentation': 'has_presentation',
            'type': 'type',
            'id': 'id'
        }

        self._label = label
        self._has_format = has_format
        self._description = description
        self._has_dimensionality = has_dimensionality
        self._has_presentation = has_presentation
        self._type = type
        self._id = id

    @classmethod
    def from_dict(cls, dikt) -> 'DataSet':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The DataSet of this DataSet.  # noqa: E501
        :rtype: DataSet
        """
        return util.deserialize_model(dikt, cls)

    @property
    def label(self):
        """Gets the label of this DataSet.


        :return: The label of this DataSet.
        :rtype: str
        """
        return self._label

    @label.setter
    def label(self, label):
        """Sets the label of this DataSet.


        :param label: The label of this DataSet.
        :type label: str
        """

        self._label = label

    @property
    def has_format(self):
        """Gets the has_format of this DataSet.


        :return: The has_format of this DataSet.
        :rtype: str
        """
        return self._has_format

    @has_format.setter
    def has_format(self, has_format):
        """Sets the has_format of this DataSet.


        :param has_format: The has_format of this DataSet.
        :type has_format: str
        """

        self._has_format = has_format

    @property
    def description(self):
        """Gets the description of this DataSet.


        :return: The description of this DataSet.
        :rtype: str
        """
        return self._description

    @description.setter
    def description(self, description):
        """Sets the description of this DataSet.


        :param description: The description of this DataSet.
        :type description: str
        """

        self._description = description

    @property
    def has_dimensionality(self):
        """Gets the has_dimensionality of this DataSet.


        :return: The has_dimensionality of this DataSet.
        :rtype: int
        """
        return self._has_dimensionality

    @has_dimensionality.setter
    def has_dimensionality(self, has_dimensionality):
        """Sets the has_dimensionality of this DataSet.


        :param has_dimensionality: The has_dimensionality of this DataSet.
        :type has_dimensionality: int
        """

        self._has_dimensionality = has_dimensionality

    @property
    def has_presentation(self):
        """Gets the has_presentation of this DataSet.


        :return: The has_presentation of this DataSet.
        :rtype: List[VariablePresentation]
        """
        return self._has_presentation

    @has_presentation.setter
    def has_presentation(self, has_presentation):
        """Sets the has_presentation of this DataSet.


        :param has_presentation: The has_presentation of this DataSet.
        :type has_presentation: List[VariablePresentation]
        """

        self._has_presentation = has_presentation

    @property
    def type(self):
        """Gets the type of this DataSet.


        :return: The type of this DataSet.
        :rtype: List[str]
        """
        return self._type

    @type.setter
    def type(self, type):
        """Sets the type of this DataSet.


        :param type: The type of this DataSet.
        :type type: List[str]
        """

        self._type = type

    @property
    def id(self):
        """Gets the id of this DataSet.


        :return: The id of this DataSet.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id):
        """Sets the id of this DataSet.


        :param id: The id of this DataSet.
        :type id: str
        """
        if id is None:
            raise ValueError("Invalid value for `id`, must not be `None`")  # noqa: E501

        self._id = id