# server/openapi_server/models/parameter.py
# coding: utf-8

from __future__ import absolute_import

from typing import List  # noqa: F401

from openapi_server import util
from openapi_server.models.base_model_ import Model


class Parameter(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, id=None, type=None, label=None, description=None, has_default_value=None, has_data_type=None):  # noqa: E501
        """Parameter - a model defined in OpenAPI

        :param id: The id of this Parameter.  # noqa: E501
        :type id: str
        :param type: The type of this Parameter.  # noqa: E501
        :type type: List[str]
        :param label: The label of this Parameter.  # noqa: E501
        :type label: str
        :param description: The description of this Parameter.  # noqa: E501
        :type description: str
        :param has_default_value: The has_default_value of this Parameter.  # noqa: E501
        :type has_default_value: str
        :param has_data_type: The has_data_type of this Parameter.  # noqa: E501
        :type has_data_type: str
        """
        self.openapi_types = {
            'id': str,
            'type': List[str],
            'label': str,
            'description': str,
            'has_default_value': str,
            'has_data_type': str
        }

        self.attribute_map = {
            'id': 'id',
            'type': 'type',
            'label': 'label',
            'description': 'description',
            'has_default_value': 'has_default_value',
            'has_data_type': 'has_data_type'
        }

        self._id = id
        self._type = type
        self._label = label
        self._description = description
        self._has_default_value = has_default_value
        self._has_data_type = has_data_type

    @classmethod
    def from_dict(cls, dikt) -> 'Parameter':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Parameter of this Parameter.  # noqa: E501
        :rtype: Parameter
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self):
        """Gets the id of this Parameter.


        :return: The id of this Parameter.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id):
        """Sets the id of this Parameter.


        :param id: The id of this Parameter.
        :type id: str
        """
        if id is None:
            raise ValueError("Invalid value for `id`, must not be `None`")  # noqa: E501

        self._id = id

    @property
    def type(self):
        """Gets the type of this Parameter.


        :return: The type of this Parameter.
        :rtype: List[str]
        """
        return self._type

    @type.setter
    def type(self, type):
        """Sets the type of this Parameter.


        :param type: The type of this Parameter.
        :type type: List[str]
        """

        self._type = type

    @property
    def label(self):
        """Gets the label of this Parameter.


        :return: The label of this Parameter.
        :rtype: str
        """
        return self._label

    @label.setter
    def label(self, label):
        """Sets the label of this Parameter.


        :param label: The label of this Parameter.
        :type label: str
        """

        self._label = label

    @property
    def description(self):
        """Gets the description of this Parameter.


        :return: The description of this Parameter.
        :rtype: str
        """
        return self._description

    @description.setter
    def description(self, description):
        """Sets the description of this Parameter.


        :param description: The description of this Parameter.
        :type description: str
        """

        self._description = description

    @property
    def has_default_value(self):
        """Gets the has_default_value of this Parameter.


        :return: The has_default_value of this Parameter.
        :rtype: str
        """
        return self._has_default_value

    @has_default_value.setter
    def has_default_value(self, has_default_value):
        """Sets the has_default_value of this Parameter.


        :param has_default_value: The has_default_value of this Parameter.
        :type has_default_value: str
        """

        self._has_default_value = has_default_value

    @property
    def has_data_type(self):
        """Gets the has_data_type of this Parameter.


        :return: The has_data_type of this Parameter.
        :rtype: str
        """
        return self._has_data_type

    @has_data_type.setter
    def has_data_type(self, has_data_type):
        """Sets the has_data_type of this Parameter.


        :param has_data_type: The has_data_type of this Parameter.
        :type has_data_type: str
        """

        self._has_data_type = has_data_type