# server/openapi_server/models/process.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model_ import Model
from openapi_server import util


class Process(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, id=None, type=None, label=None, influences=None):  # noqa: E501
        """Process - a model defined in OpenAPI

        :param id: The id of this Process.  # noqa: E501
        :type id: str
        :param type: The type of this Process.  # noqa: E501
        :type type: List[str]
        :param label: The label of this Process.  # noqa: E501
        :type label: str
        :param influences: The influences of this Process.  # noqa: E501
        :type influences: List[Process]
        """
        self.openapi_types = {
            'id': str,
            'type': List[str],
            'label': str,
            'influences': List[Process]
        }

        self.attribute_map = {
            'id': 'id',
            'type': 'type',
            'label': 'label',
            'influences': 'influences'
        }

        self._id = id
        self._type = type
        self._label = label
        self._influences = influences

    @classmethod
    def from_dict(cls, dikt) -> 'Process':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Process of this Process.  # noqa: E501
        :rtype: Process
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self):
        """Gets the id of this Process.


        :return: The id of this Process.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id):
        """Sets the id of this Process.


        :param id: The id of this Process.
        :type id: str
        """
        if id is None:
            raise ValueError("Invalid value for `id`, must not be `None`")  # noqa: E501

        self._id = id

    @property
    def type(self):
        """Gets the type of this Process.


        :return: The type of this Process.
        :rtype: List[str]
        """
        return self._type

    @type.setter
    def type(self, type):
        """Sets the type of this Process.


        :param type: The type of this Process.
        :type type: List[str]
        """

        self._type = type

    @property
    def label(self):
        """Gets the label of this Process.


        :return: The label of this Process.
        :rtype: str
        """
        return self._label

    @label.setter
    def label(self, label):
        """Sets the label of this Process.


        :param label: The label of this Process.
        :type label: str
        """

        self._label = label

    @property
    def influences(self):
        """Gets the influences of this Process.

        Property that describe the causal relationship between two processes, two variables, a process and a variable or a variable and a process  # noqa: E501

        :return: The influences of this Process.
        :rtype: List[Process]
        """
        return self._influences

    @influences.setter
    def influences(self, influences):
        """Sets the influences of this Process.

        Property that describe the causal relationship between two processes, two variables, a process and a variable or a variable and a process  # noqa: E501

        :param influences: The influences of this Process.
        :type influences: List[Process]
        """

        self._influences = influences